10ci

+++ src/main/java/life/wmm/commulity/community/dto/QuestionDTO.java
@@ -5,11 +5,11 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Long id;
+    private Integer id;
     private String title;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
